Make runtime seams explicit so host shells stop owning core assembly
The coding-grade runtime should read as a small set of explicit seams instead of hiding bootstrap and turn execution inside large host files. This extracts app-side turn bootstrap into turn_runtime.rs, splits runtime_env side effects, and moves the control-plane turn execution shell into its own helper module so chat.rs and control_plane_server.rs can stay thinner and more honest about what they own.

Constraint: Keep the single-entry product shape and 7-crate DAG unchanged
Rejected: Split new runtime/session crates now | ownership seams need to stabilize first
Confidence: high
Scope-risk: moderate
Reversibility: clean
Directive: Keep CLI and control-plane shells thin; attach future turn bootstrap or turn-submit work to the extracted seams instead of regrowing chat.rs or control_plane_server.rs
Tested: ./scripts/cargo-local-toolchain.sh fmt --all -- --check; ./scripts/cargo-local-toolchain.sh check -p loong-app; ./scripts/cargo-local-toolchain.sh test -p loong-app latest_session_selector -- --test-threads=1; ./scripts/cargo-local-toolchain.sh test -p loong-app initialize_runtime_environment_ -- --test-threads=1; ./scripts/cargo-local-toolchain.sh test -p loong-app cli_chat_options_for_turn_request_ignores_blank_working_directory -- --test-threads=1; ./scripts/cargo-local-toolchain.sh clippy -p loong-app --all-targets -- -D warnings; ./scripts/cargo-local-toolchain.sh check -p loong; ./scripts/cargo-local-toolchain.sh test -p loong turn_submit_ -- --test-threads=1; ./scripts/cargo-local-toolchain.sh clippy -p loong --all-targets -- -D warnings
Not-tested: full workspace test suite; control-plane result/stream routes outside turn_submit-filtered tests
